ILNB-backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cssselect==1.3.0
dataclasses-json==0.6.7
distro==1.9.0
dj-database-url==2.1.0
Django==4.2.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
frozendict==2.4.6
frozenlist==1.5.0
google-ai-generativelanguage==0.4.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.9.0
google-generativeai==0.3.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.63.0rc1
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jieba3k==0.35.1
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.4
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.5
langchain-text-splitters==0.3.8
langsmith==0.0.87
lxml==4.9.3
marshmallow==3.26.1
multidict==6.3.2
multitasking==0.0.11
mypy-extensions==1.0.0
newspaper3k==0.2.8
nltk==3.9.1
numpy==1.26.3
openai==1.10.0
orjson==3.10.16
packaging==23.2
pandas==2.1.4
pandas_ta==0.3.14b0
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.7
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.6
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rsa==4.9
setuptools==78.1.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
sqlparse==0.5.3
tenacity==8.5.0
tiktoken==0.5.2
tinysegmenter==0.3
tldextract==5.1.3
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
webencodings==0.5.1
websockets==15.0.1
wheel==0.45.1
whitenoise==6.6.0
yarl==1.18.3
yfinance==0.2.55
zstandard==0.23.0
